feat: Unificação de tipos no Buscador Inteligente
- Centralização de todos os tipos no arquivo globalTypes.ts, eliminando duplicações e fragmentações.
- Atualização de importações em diversos arquivos para refletir a nova estrutura.
- Remoção de arquivos de tipos obsoletos e redundantes, como types.ts, custom-express.d.ts, ncm.ts e session.ts.
- Criação de um arquivo unificado de tipos, melhorando a organização e a manutenção do código.
- Preservação de namespaces para compatibilidade com o código existente.
