+++ buscador_inteligente/src/tools/read.ts
@@ -1,8 +1,7 @@
-import https from 'https';
+import https from "https";
 import { TokenTracker } from "../utils/token-tracker";
-
-import { ReadResponse } from '../types';
-import {JINA_API_KEY} from "../config";
+import { ReadResponse } from "../types/globalTypes";
+import { JINA_API_KEY } from "../config";
 
 /**
  * Lê uma URL e retorna o conteúdo da resposta.
@@ -10,47 +9,50 @@ import {JINA_API_KEY} from "../config";
  * @param tracker Rastreador de tokens.
  * @returns O conteúdo da resposta e o número de tokens usados.
  */
-export function readUrl(url: string, tracker?: TokenTracker): Promise<{ response: ReadResponse, tokens: number }> {
+export function readUrl(
+  url: string,
+  tracker?: TokenTracker
+): Promise<{ response: ReadResponse; tokens: number }> {
   // Retorna uma promise
   return new Promise((resolve, reject) => {
     // Cria o corpo da requisição
-    const data = JSON.stringify({url});
+    const data = JSON.stringify({ url });
 
     // Cria as opções da requisição
     const options = {
       // Define o hostname
-      hostname: 'r.jina.ai',
+      hostname: "r.jina.ai",
       // Define a porta
       port: 443,
       // Define o path
-      path: '/',
+      path: "/",
       // Define o método
-      method: 'POST',
+      method: "POST",
       // Define os headers
       headers: {
         // Define o Accept
-        'Accept': 'application/json',
+        Accept: "application/json",
         // Define o Authorization
-        'Authorization': `Bearer ${JINA_API_KEY}`,
+        Authorization: `Bearer ${JINA_API_KEY}`,
         // Define o Content-Type
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
         // Define o Content-Length
-        'Content-Length': data.length,
+        "Content-Length": data.length,
         // Define o X-Retain-Images
-        'X-Retain-Images': 'none',
+        "X-Retain-Images": "none",
         // Define o X-Return-Format
-        'X-Return-Format': 'markdown'
-      }
+        "X-Return-Format": "markdown",
+      },
     };
 
     // Cria a requisição
     const req = https.request(options, (res) => {
       // Inicializa a resposta
-      let responseData = '';
+      let responseData = "";
       // Adiciona o evento de data
-      res.on('data', (chunk) => responseData += chunk);
+      res.on("data", (chunk) => (responseData += chunk));
       // Adiciona o evento de fim
-      res.on('end', () => {
+      res.on("end", () => {
         // Converte a resposta para JSON
         const response = JSON.parse(responseData) as ReadResponse;
         // Loga a resposta original
@@ -59,7 +61,7 @@ export function readUrl(url: string, tracker?: TokenTracker): Promise<{ response
         // Se o código da resposta for 402, rejeita a promise
         if (response.code === 402) {
           // Rejeita a promise
-          reject(new Error(response.readableMessage || 'Insufficient balance'));
+          reject(new Error(response.readableMessage || "Insufficient balance"));
           // Retorna
           return;
         }
@@ -67,32 +69,32 @@ export function readUrl(url: string, tracker?: TokenTracker): Promise<{ response
         // Se a resposta não contém dados, rejeita a promise
         if (!response.data) {
           // Rejeita a promise
-          reject(new Error('Invalid response data'));
+          reject(new Error("Invalid response data"));
           // Retorna
           return;
         }
 
         // Loga a resposta
-        console.log('Read:', {
+        console.log("Read:", {
           // Título
           title: response.data.title,
           // URL
           url: response.data.url,
           // Tokens
-          tokens: response.data.usage?.tokens || 0
+          tokens: response.data.usage?.tokens || 0,
         });
 
         // Obtém o número de tokens
         const tokens = response.data.usage?.tokens || 0;
         // Rastrea o uso de tokens
-        (tracker || new TokenTracker()).trackUsage('read', tokens);
+        (tracker || new TokenTracker()).trackUsage("read", tokens);
         // Resolve a promise
         resolve({ response, tokens });
       });
     });
 
     // Adiciona o evento de erro
-    req.on('error', reject);
+    req.on("error", reject);
     // Escreve o corpo da requisição
     req.write(data);
     // Finaliza a requisição
