--- buscador_inteligente/src/types.ts
@@ -1,357 +1,0 @@
-import { CoreAssistantMessage, CoreUserMessage, LanguageModelUsage } from "ai";
-import { TokenTracker } from "./utils/token-tracker";
-import { ActionTracker } from "./utils/action-tracker";
-import { Schema as GoogleSchema, SchemaType } from "@google/generative-ai";
-
-export type SchemaProperty = {
-  type: SchemaType;
-  description?: string;
-  items?: SchemaProperty;
-  properties?: Record<string, SchemaProperty>;
-  required?: string[];
-  enum?: string[];
-  maxItems?: number;
-  minItems?: number;
-};
-
-// Re-exportando os tipos da biblioteca
-export { SchemaType };
-export type Schema = GoogleSchema;
-
-// Definição correta de ResponseSchema para ser compatível com @google/generative-ai
-export type ResponseSchema = {
-  type: SchemaType.OBJECT;
-  properties: Record<string, Schema>;
-  required?: string[];
-};
-
-// Tipos de Bloco de Conteúdo
-export interface ContentBlock {
-  type: string;
-  content: string;
-}
-
-export interface TextBlock extends ContentBlock {
-  type: "text";
-}
-
-export interface CodeBlock extends ContentBlock {
-  type: "code";
-  language?: string;
-}
-
-export interface HeadingBlock extends ContentBlock {
-  type: "heading";
-  level: number;
-}
-
-export interface ListItem {
-  content: string;
-  items?: ListItem[];
-}
-
-export interface ListOrderedBlock extends ContentBlock {
-  type: "list-ordered";
-  items: ListItem[];
-}
-
-export interface ListUnorderedBlock extends ContentBlock {
-  type: "list-unordered";
-  items: ListItem[];
-}
-
-export interface TableBlock extends ContentBlock {
-  type: "table";
-  headers: string[];
-  rows: string[][];
-}
-
-export interface QuoteBlock extends ContentBlock {
-  type: "quote";
-}
-
-export interface AlertBlock extends ContentBlock {
-  type: "alert";
-  variant: "info" | "warning" | "error" | "success";
-}
-
-// Tipos de Ação
-type BaseAction = {
-  action: "search" | "answer" | "reflect" | "visit" | "coding";
-  think: string;
-  accumulatedReasoning?: string;
-};
-
-export type SearchAction = BaseAction & {
-  action: "search";
-  searchQuery: string;
-  searchResults?: SearchResult[] | QueryResult[];
-};
-
-export type AnswerAction = BaseAction & {
-  action: "answer";
-  answer: string;
-  references: Array<{
-    exactQuote: string;
-    url: string;
-  }>;
-  isFinal?: boolean;
-  reasoning?: string;
-};
-
-export type KnowledgeItem = {
-  question: string;
-  answer: string;
-  references?:
-    | Array<{
-        exactQuote: string;
-        url: string;
-      }>
-    | Array<any>;
-  type: "qa" | "side-info" | "chat-history" | "url" | "coding";
-  updated: string;
-};
-
-export type ReflectAction = BaseAction & {
-  action: "reflect";
-  questionsToAnswer: string[];
-};
-
-export type VisitAction = BaseAction & {
-  action: "visit";
-  URLTargets: string[];
-};
-
-export type CodingAction = BaseAction & {
-  action: "coding";
-  codingIssue: string;
-};
-
-export type StepAction =
-  | SearchAction
-  | AnswerAction
-  | ReflectAction
-  | VisitAction
-  | CodingAction;
-
-export type EvaluationType =
-  | "definitive"
-  | "freshness"
-  | "plurality"
-  | "attribution";
-export type EvaluationCriteria = {
-  types: EvaluationType[];
-  languageStyle: string;
-  maxAgeDays?: number;
-  think?: string;
-  pass?: boolean;
-  tokens?: number;
-};
-
-// Interface de Uso de Tokens
-export interface TokenUsage {
-  tool: string;
-  tokens: number;
-  usage?: LanguageModelUsage;
-}
-
-// Outras interfaces e tipos
-export interface SearchResponse {
-  code: number;
-  status: number;
-  data: Array<{
-    title: string;
-    description: string;
-    url: string;
-    content: string;
-    usage: { tokens: number };
-  }> | null;
-  name?: string;
-  message?: string;
-  readableMessage?: string;
-}
-
-export interface BraveSearchResponse {
-  web: {
-    results: Array<{
-      title: string;
-      description: string;
-      url: string;
-    }>;
-  };
-}
-
-export type DedupResponse = {
-  think: string;
-  unique_queries: string[];
-};
-
-export interface ReadResponse {
-  code: number;
-  status: number;
-  data?: {
-    title: string;
-    description: string;
-    url: string;
-    content: string;
-    usage: { tokens: number };
-  };
-  name?: string;
-  message?: string;
-  readableMessage?: string;
-}
-
-export type EvaluationResponse = {
-  pass: boolean;
-  think: string;
-  tokens?: number;
-  type?: "definitive" | "freshness" | "plurality" | "attribution";
-  freshness_analysis?: {
-    likely_outdated: boolean;
-    dates_mentioned: string[];
-    current_time: string;
-    max_age_days?: number;
-  };
-  plurality_analysis?: {
-    expects_multiple: boolean;
-    provides_multiple: boolean;
-    count_expected?: number;
-    count_provided: number;
-  };
-  attribution_analysis?: {
-    sources_provided: boolean;
-    sources_verified: boolean;
-    quotes_accurate: boolean;
-  };
-};
-
-export type ErrorAnalysisResponse = {
-  recap?: string;
-  blame?: string;
-  improvement?: string;
-  questionsToAnswer?: string[];
-};
-
-export interface SearchResult {
-  title: string;
-  url: string;
-  description: string;
-}
-
-export interface QueryResult {
-  query: string;
-  results: SearchResult[];
-}
-
-export interface StepData {
-  step: number;
-  question: string;
-  action: string;
-  reasoning: string;
-  searchQuery?: string;
-  result?: QueryResult[];
-}
-
-export type KeywordsResponse = {
-  think: string;
-  queries: string[];
-};
-
-export interface StreamMessage {
-  trackers: {
-    tokenTracker: TokenTracker;
-    actionTracker: ActionTracker;
-  };
-  type:
-    | "progress"
-    | "answer"
-    | "error"
-    | "search"
-    | "reflect"
-    | "visit"
-    | "log"
-    | "connected";
-  data: string | StepAction;
-  outputs?: any[];
-  step?: number;
-  budget?: {
-    used: number;
-    total: number;
-    percentage: string;
-  };
-}
-
-// Tipos da API OpenAI
-export interface Model {
-  id: string;
-  object: "model";
-  created: number;
-  owned_by: string;
-}
-
-export interface ChatCompletionRequest {
-  model: string;
-  messages: Array<CoreUserMessage | CoreAssistantMessage>;
-  stream?: boolean;
-  reasoning_effort?: "low" | "medium" | "high" | null;
-  max_completion_tokens?: number | null;
-}
-
-export interface ChatCompletionResponse {
-  id: string;
-  object: "chat.completion";
-  created: number;
-  model: string;
-  system_fingerprint: string;
-  choices: Array<{
-    index: number;
-    message: {
-      role: "assistant";
-      content: string;
-    };
-    logprobs: null;
-    finish_reason: "stop";
-  }>;
-  usage: {
-    prompt_tokens: number;
-    completion_tokens: number;
-    total_tokens: number;
-  };
-}
-
-export interface ChatCompletionChunk {
-  id: string;
-  object: "chat.completion.chunk";
-  created: number;
-  model: string;
-  system_fingerprint: string;
-  choices: Array<{
-    index: number;
-    delta: {
-      role?: "assistant";
-      content?: string;
-    };
-    logprobs: null;
-    finish_reason: null | "stop";
-  }>;
-  usage?: any;
-}
-
-// Contexto do Rastreador
-export interface TrackerContext {
-  outputs: any[];
-  tokenTracker: TokenTracker;
-  actionTracker: ActionTracker;
-}
-
-// Atualização da interface ServerLog para incluir requestId
-export interface ServerLog {
-  context: {
-    pid: number;
-    env: string;
-    requestId?: string;
-  };
-  timestamp: string;
-  message: string;
-  level: "log" | "error" | "warn" | "info";
-}
