--- buscador_inteligente/src/types/custom-express.d.ts
@@ -1,9 +1,0 @@
-import { FastApiNCMResult } from "./ncm";
-
-declare global {
-  namespace Express {
-    interface Request {
-      fastApiResult?: FastApiNCMResult | null;
-    }
-  }
-}
