--- buscador_inteligente/src/types/index 2.ts
@@ -1,10 +1,0 @@
-// Exportação de tipos globais
-export * from './globalTypes';
-
-// Importação dos módulos
-import * as TestTypesInternal from './TestTypes';
-import * as SessionModule from './session';
-
-// Re-exportação com namespace
-export const TestModule = TestTypesInternal;
-export { SessionModule };
