+++ buscador_inteligente/src/types/unictypes.md
@@ -1,0 +1,69 @@
+# Unificação de Tipos do Buscador Inteligente
+
+## Implementação de Estratégia de Tipos Unificados
+
+### Descrição da Mudança
+
+Implementamos uma estratégia de tipos totalmente unificados para o Buscador Inteligente. Todos os tipos foram centralizados em um único arquivo `globalTypes.ts`, eliminando a fragmentação e duplicação de definições de tipos.
+
+### Arquivos Afetados
+
+- `buscador_inteligente/src/types/index.ts`: Atualizado para exportar apenas os tipos de globalTypes.ts
+- `buscador_inteligente/src/types/globalTypes.ts`: Contém agora todos os tipos unificados do projeto
+- `buscador_inteligente/src/types/express.d.ts`: Atualizado para importar de globalTypes.ts
+- Diversos arquivos com importações corrigidas para apontar para `types/globalTypes.ts`
+
+### Arquivos Removidos
+
+- `buscador_inteligente/src/types.ts`: Arquivo de tipos raiz que foi substituído pela pasta `types/`
+- `buscador_inteligente/src/types/custom-express.d.ts`: Removido por ser duplicação de `express.d.ts`
+- `buscador_inteligente/src/types/index 2.ts`: Removido por ser uma versão redundante
+- `buscador_inteligente/src/types/ncm.ts`: Todo conteúdo movido para globalTypes.ts
+- `buscador_inteligente/src/types/session.ts`: Todo conteúdo movido para globalTypes.ts e namespace em index.ts
+
+### Otimizações Realizadas
+
+1. **Total Centralização**: Todos os tipos estão agora em um único arquivo
+2. **Eliminação de Duplicações**: Tipos redundantes foram unificados
+3. **Organização Lógica**: Tipos estão organizados e documentados por seção
+4. **Namespaces Preservados**: O namespace SessionModule foi preservado para manter compatibilidade
+5. **Padronização de Importações**: Todas as importações apontam agora para a pasta `types/` ao invés de arquivos individuais
+
+### Tipos Adicionados ou Transferidos
+
+Os seguintes tipos foram transferidos do arquivo `src/types.ts` para `src/types/globalTypes.ts`:
+
+- Tipos de Bloco de Conteúdo: `ContentBlock`, `TextBlock`, `CodeBlock`, `HeadingBlock`, etc.
+- Interface de Log do Servidor: `ServerLog`
+- Outros tipos específicos do sistema
+
+### Compatibilidade Mantida
+
+Para garantir que o código existente continue funcionando sem alterações:
+
+- Mantivemos o namespace `SessionModule` no arquivo index.ts
+- Todas as exportações originais continuam disponíveis com os mesmos nomes
+- Todas as importações internas foram atualizadas para refletir a nova estrutura
+
+### Guia de Uso
+
+Para usar os tipos no projeto:
+
+```typescript
+// Importação de todos os tipos
+import * as Types from '../types';
+
+// Importação de tipos específicos
+import { ConsultaProduto, QueryStep } from '../types';
+
+// Importação de tipos que antes estavam em namespaces
+import { SessionModule } from '../types';
+const session: SessionModule.QuerySession = { ... };
+```
+
+### Vantagens da Nova Estrutura
+
+- **Manutenção Simplificada**: Um único arquivo para editar todos os tipos
+- **Consistência Garantida**: Eliminação de definições duplicadas ou conflitantes
+- **Melhor Documentação**: Tipos agrupados logicamente e bem documentados
+- **Performance Melhorada**: Menos arquivos para o TypeScript processar
