+++ shared-types/src/query/index.ts
@@ -1,8 +1,8 @@
 // Exportando todos os tipos
-export * from './types';
+export * from "./types";
 
 // Exportando todos os schemas
-export * from './schemas';
+export * from "./schemas";
 
 // Exportando todos os transformadores
-export * from './transformers'; 
+export * from "./transformers";
